Add asset pack toggle
This creates a new top-level store entry "assetPack" and uses it to
track which asset pack is currently in use. Right now these values are
simply hard-coded, but they're laid on in a predictable format on disk.

Since not all assets are supported currently we use a JSON index to
track which files exist, avoiding an extraneous network roundtrip.