Update fr.json (#84)
* Update fr.json

This is my very first time "contributing" to a project on GitHub, and I hope I'm doing everything correctly! I noticed a small typo in the species name in the French translation file. The common name was listed as "Vautour des palombes," but the correct term is "Autour des Palombes" (Northern Goshawk). I've made this change to improve accuracy. Thank you for reviewing my contribution!

* Update typo bird 284 in the FR version

* Publish changes